Refactor: Change user weight data type from Float to Double
This commit refactors the application to use `Double` instead of `Float` for handling user weight, providing better precision. The `userWeight` field in `SignInResult` is now a nullable `Double` to better represent cases where the weight is not set.

Additionally, the `OnBoardingRepository` has been improved to only save non-null and non-default user data fields to Firestore, preventing empty values from being written.

Key changes:
- **Data Types and Models**:
  - **`SignInResult.kt`**: Changed `userWeight` from `Float` to `Double?`.
  - **`Preferences.kt`**: Updated `getWeight()` and `setWeight()` to handle `Double`, with conversions to and from `Float` for storage.
  - **`CaloriesBurntCalculator.kt`**: The `calculatePushupCalories` function now accepts `weightKg` as a `Double`.

- **Repositories**:
  - **`OnBoardingRepository.kt`**: Now selectively saves only non-empty user data fields to Firestore, avoiding null or default values.
  - **`AccountRepository.kt`** & **`WeightSetupRepository.kt`**: Updated methods (`updateWeight`, `saveData`) to accept a `Double` for the user's weight.

- **UI and ViewModels**:
  - Across `AccountFragment`, `AccountViewModel`, `WeightSetupFragment`, and `WeightSetupViewModel`, the logic has been updated to use `Double` for weight-related operations and data handling.
  - **`WeightSetupFragment.kt`**: Added logic to automatically navigate to the next screen if weight data is already saved.

+++ app/src/main/java/com/subhajitrajak/durare/data/repositories/AccountRepository.kt
@@ -16,7 +16,7 @@ class AccountRepository (val context: Context) {
     val userDoc = uid?.let { db.collection(Constants.USERS).document(it) }
     val profileDoc = userDoc?.collection(Constants.USER_DATA)?.document(Constants.PROFILE)
 
-    suspend fun updateWeight(weight: Float) {
+    suspend fun updateWeight(weight: Double) {
         pref.setWeight(weight)
         profileDoc?.update(Constants.USER_WEIGHT, weight)?.await()
     }
