+++ app/src/main/java/com/subhajitrajak/durare/utils/CaloriesBurntCalculator.kt
@@ -3,7 +3,7 @@ package com.subhajitrajak.durare.utils
 import java.util.Locale
 
 fun calculatePushupCalories(
-    weightKg: Float,           // user's weight in kilograms
+    weightKg: Double,           // user's weight in kilograms
     pushupCount: Int,           // total number of pushups done
     pushupsPerMinute: Float, // average pushup rate (default 25/min)
     metValue: Float = 8.0f      // MET value for pushups (moderate to vigorous)
